+++ packages/editor/renderer/src/__tests__/unit/react/nodes/inlineExtension.tsx
@@ -1,0 +1,172 @@
+import * as React from 'react';
+import { mount } from 'enzyme';
+import { expect } from 'chai';
+import InlineExtension from '../../../../react/nodes/inlineExtension';
+import { RendererContext } from '../../../../react';
+import ReactSerializer from '../../../../react';
+import { defaultSchema, ExtensionHandlers } from '@atlaskit/editor-common';
+
+describe('Renderer - React/Nodes/InlineExtension', () => {
+  const extensionHandlers: ExtensionHandlers = {
+    'com.atlassian.fabric': (param: any, doc: any) => {
+      switch (param.extensionKey) {
+        case 'react':
+          return <span>This is a react element</span>;
+        case 'adf':
+          return [
+            {
+              type: 'text',
+              text: 'This is a ADF node',
+            },
+          ];
+        case 'error':
+          throw new Error('Cursed by Tong');
+        default:
+          return null;
+      }
+    },
+  };
+
+  const rendererContext: RendererContext = {
+    adDoc: {
+      version: 1,
+      type: 'doc',
+      content: [
+        {
+          type: 'paragraph',
+          content: [
+            {
+              type: 'text',
+              text: 'Check out this extension',
+            },
+          ],
+        },
+        {
+          type: 'inlineExtension',
+          attrs: {
+            extensionType: 'com.atlassian.stride',
+            extensionKey: 'default',
+            bodyType: 'none',
+          },
+          content: [
+            {
+              type: 'text',
+              text: 'This is the default content of the extension',
+            },
+          ],
+        },
+      ],
+    },
+    schema: defaultSchema,
+  };
+
+  const serializer = new ReactSerializer({});
+
+  it('should be able to fall back to default content', () => {
+    const extension = mount(
+      <InlineExtension
+        serializer={serializer}
+        extensionHandlers={extensionHandlers}
+        rendererContext={rendererContext}
+        extensionType="com.atlassian.fabric"
+        extensionKey="default"
+        text="This is the default text"
+      />,
+    );
+
+    expect(
+      extension
+        .find('span')
+        .first()
+        .text(),
+    ).to.equal('This is the default text');
+    extension.unmount();
+  });
+
+  it('should be able to render React.Element from extensionHandler', () => {
+    const extension = mount(
+      <InlineExtension
+        serializer={serializer}
+        extensionHandlers={extensionHandlers}
+        rendererContext={rendererContext}
+        extensionType="com.atlassian.fabric"
+        extensionKey="react"
+      />,
+    );
+
+    expect(
+      extension
+        .find('span')
+        .first()
+        .text(),
+    ).to.equal('This is a react element');
+    extension.unmount();
+  });
+
+  it('should be able to render Atlassian Document from extensionHandler', () => {
+    const extension = mount(
+      <InlineExtension
+        serializer={serializer}
+        extensionHandlers={extensionHandlers}
+        rendererContext={rendererContext}
+        extensionType="com.atlassian.fabric"
+        extensionKey="adf"
+      />,
+    );
+
+    expect(
+      extension
+        .find('span')
+        .first()
+        .text(),
+    ).to.equal('This is a ADF node');
+    extension.unmount();
+  });
+
+  it('should render the default content if extensionHandler throws an exception', () => {
+    const extension = mount(
+      <InlineExtension
+        serializer={serializer}
+        extensionHandlers={extensionHandlers}
+        rendererContext={rendererContext}
+        extensionType="com.atlassian.fabric"
+        extensionKey="error"
+      />,
+    );
+
+    expect(
+      extension
+        .find('span')
+        .first()
+        .text(),
+    ).to.equal('inlineExtension');
+    extension.unmount();
+  });
+
+  it('extension handler should receive type = inlineExtension', () => {
+    const extensionHandler = jest.fn();
+    const extensionHandlers: ExtensionHandlers = {
+      'com.atlassian.fabric': extensionHandler,
+    };
+
+    const extension = mount(
+      <InlineExtension
+        serializer={serializer}
+        extensionHandlers={extensionHandlers}
+        rendererContext={rendererContext}
+        extensionType="com.atlassian.fabric"
+        extensionKey="react"
+      />,
+    );
+
+    expect(extensionHandler.mock.calls[0][0]).to.eql({
+      type: 'inlineExtension',
+      extensionType: 'com.atlassian.fabric',
+      extensionKey: 'react',
+      parameters: undefined,
+      content: undefined,
+    });
+
+    extension.unmount();
+  });
+});