+++ packages/elements/analytics-listeners/src/types.ts
@@ -26,7 +26,7 @@ export type ListenerFunction = (event: EventNextType) => void;
 
 export enum FabricChannel {
   atlaskit = 'atlaskit',
-  elements = 'elements',
+  elements = 'fabric-elements',
   navigation = 'navigation',
   editor = 'editor',
 }