+++ packages/search/global-search/src/__tests__/unit/common/SearchResults.test.tsx
@@ -16,7 +16,7 @@ const defaultProps = {
   retrySearch: () => {},
   keepPreQueryState: false,
   searchSessionId: 'session_id',
-  renderAdvancedSearchLink: () => <div id="advanced-search-link" />,
+  renderNoRecentActivity: () => <div id="advanced-search-link" />,
   renderNoResult: () => <div id="no-result" />,
   renderAdvancedSearchGroup: () => <div id="advanced-search-group" />,
   getPreQueryGroups: () => [],
@@ -82,7 +82,7 @@ it('should render pre query state when there is no query entered', () => {
   expect(preQueryState.length).toBe(1);
   expect(preQueryState.props()).toMatchObject({
     query: '',
-    renderAdvancedSearchLink: defaultProps.renderAdvancedSearchLink,
+    renderNoRecentActivity: defaultProps.renderNoRecentActivity,
     searchSessionId: defaultProps.searchSessionId,
     renderAdvancedSearchGroup: defaultProps.renderAdvancedSearchGroup,
     resultsGroups: mockResultsGroup,