+++ packages/search/global-search/src/components/SearchResultsUtil.tsx
@@ -14,6 +14,10 @@ export function getConfluenceAdvancedSearchLink(query?: string) {
   return `/wiki/dosearchsite.action${queryString}`;
 }
 
+export function getJiraAdvancedSearchUrl(entityType: string, query?: string) {
+  return `/search/${entityType}?query=${query}`;
+}
+
 export function redirectToConfluenceAdvancedSearch(query = '') {
   // XPSRCH-891: this breaks SPA navigation. Consumer needs to pass in a redirect/navigate function.
   window.location.assign(getConfluenceAdvancedSearchLink(query));